+++ french_locator_filter/resources/i18n/french_locator_filter_fr.ts
@@ -4,9 +4,19 @@
 <context>
     <name>FrenchBanGeocoderLocatorFilter</name>
     <message>
-        <location filename="../../core/locator_filter.py" line="67"/>
+        <location filename="../../core/locator_filter.py" line="85"/>
         <source>French Adress geocoder</source>
         <translation>Géocodeur API BAN</translation>
+    </message>
+    <message>
+        <location filename="../../core/locator_filter.py" line="169"/>
+        <source>Result triggerred by the user received: {}</source>
+        <translation>Résultat sélectionné par l&apos;utilisateur reçu : {}</translation>
+    </message>
+    <message>
+        <location filename="../../core/locator_filter.py" line="178"/>
+        <source>Something went wrong during result deserialization: {}. Trying another method...</source>
+        <translation>Impossible de récupérer le résultat sélection : {}. Essai par une autre méthode...</translation>
     </message>
 </context>
 <context>
@@ -22,64 +32,74 @@
         <translation>Fonctionnalités</translation>
     </message>
     <message>
-        <location filename="../../gui/dlg_settings.ui" line="216"/>
+        <location filename="../../gui/dlg_settings.ui" line="260"/>
         <source>HTTP user-agent:</source>
         <translation>HTTP user-agent :</translation>
     </message>
     <message>
-        <location filename="../../gui/dlg_settings.ui" line="106"/>
+        <location filename="../../gui/dlg_settings.ui" line="175"/>
         <source>Request URL:</source>
         <translation>Base de l&apos;URL de requête :</translation>
     </message>
     <message>
-        <location filename="../../gui/dlg_settings.ui" line="144"/>
+        <location filename="../../gui/dlg_settings.ui" line="125"/>
         <source>Request parameters:</source>
         <translation>Paramètres de la requête :</translation>
     </message>
     <message>
-        <location filename="../../gui/dlg_settings.ui" line="292"/>
+        <location filename="../../gui/dlg_settings.ui" line="306"/>
         <source>Miscellaneous</source>
         <translation>Divers</translation>
     </message>
     <message>
-        <location filename="../../gui/dlg_settings.ui" line="301"/>
+        <location filename="../../gui/dlg_settings.ui" line="315"/>
         <source>Enable debug mode.</source>
         <translation>Activer le mode debug.</translation>
     </message>
     <message>
-        <location filename="../../gui/dlg_settings.ui" line="310"/>
+        <location filename="../../gui/dlg_settings.ui" line="324"/>
         <source>Debug mode (degraded performances)</source>
         <translation>Activer le mode debug (performances dégradées)</translation>
     </message>
     <message>
-        <location filename="../../gui/dlg_settings.ui" line="330"/>
+        <location filename="../../gui/dlg_settings.ui" line="366"/>
         <source>Version used to save settings:</source>
         <translation>Paramètres sauvegardés avec la version :</translation>
     </message>
     <message>
-        <location filename="../../gui/dlg_settings.ui" line="349"/>
+        <location filename="../../gui/dlg_settings.ui" line="385"/>
         <source>Report an issue</source>
         <translation>Signaler une anomalie</translation>
     </message>
     <message>
-        <location filename="../../gui/dlg_settings.ui" line="360"/>
+        <location filename="../../gui/dlg_settings.ui" line="359"/>
         <source>Help</source>
         <translation>Aide en ligne</translation>
     </message>
     <message>
-        <location filename="../../gui/dlg_settings.ui" line="185"/>
+        <location filename="../../gui/dlg_settings.ui" line="144"/>
         <source>Minimal search length</source>
         <translation>Nombre minimal de caractères avant de déclencher la requête à l&apos;API</translation>
     </message>
     <message>
-        <location filename="../../gui/dlg_settings.ui" line="191"/>
+        <location filename="../../gui/dlg_settings.ui" line="150"/>
         <source> characters</source>
         <translation> caractères</translation>
     </message>
     <message>
-        <location filename="../../gui/dlg_settings.ui" line="125"/>
+        <location filename="../../gui/dlg_settings.ui" line="106"/>
         <source>Minimal search length:</source>
         <translation>Longueur minimale de la recherche :</translation>
+    </message>
+    <message>
+        <location filename="../../gui/dlg_settings.ui" line="267"/>
+        <source>HTTP content type:</source>
+        <translation>Type de contenu HTTP :</translation>
+    </message>
+    <message>
+        <location filename="../../gui/dlg_settings.ui" line="356"/>
+        <source>Open documentation</source>
+        <translation>Ouvrir l&apos;aide en ligne du plugin sur votre navigateur</translation>
     </message>
 </context>
 </TS>