+++ CHANGELOG.md
@@ -5,8 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## 1.0.4 - 2022-11-15
+## 1.0.4 - 2022-10-20
 
+- Clean legacy code
+- Fix a bug that can lead to a QGIS crash because of legacy method
 - add a button to reset settings to factory default
 
 ## 1.0.3 - 2022-10-18